Bug fixes and functionality enhancements
General:
1) Improved the appearance of logging messages via cout
(#define LOGGING).
2) All reinterpret_casts are replaced with static_casts.
3) Added showList(const List<T, Allocator>& list); method
to display the contents of the list.

For the base node:
1) Added alias NodePtr_t to define Node* and const Node*.
Used in the * and -> operators.
2) Added operator !=.

For the list:
1) Changed the name of the Alloc template parameter to
Allocator. instead of NodeAlloc now NodeAllocator
2) Methods added:
    get_allocator()
    insert (const_iterator pos, const T& value)
    insert( const_iterator pos, size_type count, const T& value )
    erase (const_iterator pos)
    erase (const_iterator first, const_iterator last)

3) Some minor changes were made to the push_back and push_front methods: the
calls std::allocator_traits<NodeAllocator>::deallocate(alloc_, new_node, 1)
was moved from the nested catch block to the outer one
4) Some minor changes were made to the pop_back and pop_front methods: now
the pointer to the removed node has the type BaseNode* instead of Node*
(since in these methods there is no need to access the T field Node::value)

For the structure A:
1) The operator<< function was added to the original A objects