Support move semantics and transition to hpp file
General
1) The implementation of the List class was moved from the executable file
"list.cpp" to the header file "list.hpp"
Accordingly, the implementation of the structure A, which simulates the values
for the list, and the temporary implementation of the allocator were moved to
the main.cpp file to which the file "Containers/include/list.hpp" is connected

For the list (List):
1) Support for c++14 and older was removed. Now the minimum version is c++17;
2) In all methods, unnecessary nesting of try{...}/catch(...){throw;} was
removed since they did not perform any useful action.
3) All references to std::allocator_traits<Allocator> were replaced with
std::allocator_traits<NodeAllocator> to use the allocator specifically for nodes.
3) Support for move semantics has been implemented, namely:
- List(List&& other);
- List& operator=(List&& other);
- terator insert(const_iterator pos, T&& value);
- void push_back(T&& value);
- reference emplace_back(Args&&... args);
- void push_front(T&& value);
- reference emplace_front(Args&&... args);
4) Corrections have been made to the logic of the copy assignment operator
List& operator=(coonst List& other):
4.1) The "if constexpr (std::allocator_traits<NodeAllocator>::is_always_equal::value)"
check has been added to optimize performance in the case of constant equality of allocators.
4.2) The copy assignment operator now provides strong guarantees. exception safety, thanks
to the Copy-and-swap approach.
5) In the assign methods:
- void assign(size_t count, const T& value);
- void assign(InputIt first, InputIt last);
the logic of replacing an existing value with a new one has been redesigned: Previously, a
complete re-creation of the value was performed, through a call to the destructor and constructor.
Now the copy assignment operator is used.
6) In the void clear() method, the erase(cbegin(), cend()); call is now used instead of the
paired call to pop_front() and pop_back();
7) In the insert methods, an insertion position (pos) error has been fixed: previously, new
elements were inserted after position (pos). Now elements are inserted before position (pos).
8) The condition was added to the void swap(List& other) noexcept{...} method:
noexcept(noexcept(std::allocator_traits<NodeAllocator>::is_always_equal::value)); Also, due to the
rejection of support for C++14 and older, the old form of working with allocator_traits was removed;
9) The following methods were added:
- size_type remove(const T& value);
- size_type remove_if( UnaryPred p );

For the iterator (base_iterator):
1) The unnecessary aliases using pointer_type_ and using refernece_type_ were removed. Accordingly,
all references to pointer_type_ and refernece_type_ were replaced with pointer and refernece