Adding second allocator and shrink_to_fit rework
1) A new allocator (AllocatorPtrOnBucket alloc_ptr_on_bucket_) has been
added for allocating pointers of type (T*) in the same memory as objects
of type (T).

2) The Shrink_to_fit() method has been reworked: it now complies with the
strict exclusion guarantee when outputs other than the move constructor
of a non-CopyInsertable T are provided
(see cppreference -> std::deque<T,Allocator>::shrink_to_fit for details).
    Previously, this method executed without realocation of objects of type T
to other memory (move was performed exclusively within the original deque
itself). Because of this, it was impossible to comply with the strong
exception guarantee (at most, basic exception guarantee could be
implemented). Now all deque is reallocated (which is permitted by the
standard -> which allows to invalidate all iterators and
references/pointers on deque elements).
    !!! A move operation from the source deque to the new (reallocated)
is using the std::move_if_noexcep() method. This means that if the
move-constructor of your type T is NOT market as noexcep, the move
operations will be replaced by copy operations.

3) The method
    iterator insertion (const_iterator pos, size_type count, const T& value)
was removed because it requires significant thought.

4) The method
    void push_back(T&& value){} was removed because it requires significant
thought.

!!!5) The methods
erase iterator (first const_iterator, then const_iterator)
    and
void push_back (const T& value)
have mediocre implementations and require significant improvements.