Fix: base_iterator, begin(), end() and erase (first, last)
1) base_iterator:
1.1) fixed operator--():
1.1.1) now the distance from the element pointer (ptr_) to the initial
bucket (*bucket_ptr_) is not used to determine the need to switch to
the previous bucket, and now a direct comparison of the element pointer
(ptr_) with the initial bucket (*bucket_ptr_) is performed
1.1.2) fixed switching to the previous bucket: now it points exactly to
the end element of the new (previous) bucket.
1.2) Made a small renaming of variables in operator-=(difference_type value)

2) begin() and end() – the constant versions of these methods now have
their own implementation. Previously, they depended on non-const versions,
but compilers warned about recursive calls despite casting (this) to a
non-const value (for explicitly calling the non-const version of method
in the const version).

3) erase(first, last): in the worst block:
3.1) the condition that determined the direction of shifting garbage
elements (to the beginning or end) was fixed
3.2) the move operation was replaced by std::swap
3.3) a bug that prevented the removal of 1 garbage element was fixed.