New template param & fix of end, erase, push_back
1) Added 3rd parameter of the BucketSize template, which determines
the size of one bucket. By default, it is 4096 bytes or 16 elements
(if 16 elements exceed 4096 bytes).
   Accordingly, an alias was added to determine the size of the bucket
using the 2nd parameter of the template (the allocator becomes the
3rd parameter of the template and also has the default value
std::allocator<T>).

2) Fixed the iterator end() method and its const version: now, if
the deque is empty, the iterator on nullptr is returned.

3) In the erase(first, last) method, in the case of optimization due
to moving whole buckets, an error was fixed, due to which the iterator
to the first garbage element was updated incorrectly.

4) in the push_back method: fixed the calculation of the base value of
the old capacity which is used to calculate the new capacity (the new
capacity is the base value of the old capacity multiplied by 3)

5) fixed some warnings caused by comparing different types.