Fix of base_iterator & rbegin(), crbegin(), rend(), crend()
1) Added handling of uninitialized deque situation in base_iterator
(deque not initialized or shrink_to_fit called for empty deque).
That is, now performing manipulations on the obtained iterators using
begin()/end() methods and similar const/reverse versions on an
uninitialized deque is safe and does not cause an abort() call
(although formally such operations are still UB).

2) Fixed methods rbegin(), crbegin(), rend(), crend() – now they return
iterators to the required elements and do not cause a warnings from
compiler.