Fix pointer constancy in base_iterator
At the beginning of this branch (2 commits ago) changes were made
regarding the constness of pointers in the constant iterator.
As a result, the constant iterator (base_terator<true>) received
changes in the type of pointers to the bucket and data:
(const T**, const T*) - now these are pointers to a constant.

But it was discovered that the constness of the bucket pointer
(const T**) does not damage modifications through this pointer
to the pointer to date:
    const T* ptr_ = ...;
    const T** bucket_ptr_ = ...;
    *bucket_ptr_ = ...; // -> OK

Therefore, it was decided to leave the bucket pointer as a simple
type (T**) without any const qualifiers.