fix/deque: fix of insert(pos, count, value)
There was a problem: after calling method insert(pos, count, value),
in the uninitialized deque (m_buckets_ptr == nullptr),the iterator
m_last pointed not to the last element, but to the position after it
(in fact, m_last was equal to the result of calling end()).

This has been fixed and now the iterator m_last points to the last
element of the deque.