fix/deque/:base_iterator operator++ & operator--
There was a problem:
When the iterator went beyond the outer deque array, its bucket pointer
(m_bucket_ptr) began to point to someone else's memory, which is why any
attempt to move the iterator further away from the deque resulted in a
UB --> dereference of the pointer to someone else's memory. (sometimes
the transition of the iterator shift is accompanied by a transition to
another bucket, which requires the execution of the dereferencing
operations:
    ++m_bucket_ptr;
    m_ptr = *m_bucket_ptr
or
    --m_bucket_ptr;
    m_ptr = *m_bucket_ptr + BucketSize + 1

This problem manifests itself precisely when the iterator goes not beyond
the allocated buckets (going beyond the limits of
[m_first_allocated_bucket_ptr, m_last_allocated_bucket_ptr]),
but specifically beyond the limits of the external array of pointers to
buckets --> [m_buckets_ptr; buckets_ptr + m_buckets_capacity].

This problem was manifested only by sanitizers (without them, the UB did
not affect the operation of the program in any way).

Now, this problem is solved: now the iterator knows the boundaries of the
deque (it has a pointer to the external array of pointers m_buckets_ptr
and its capacity m_buckets_capacity) and it also has an internal pseudo-index
(m_pseudo_cell_index) which replaces the data pointer (m_ptr) while the
iterator finds outside the deque.
This helps to avoid operations (m_ptr = *m_bucket_ptr) or
(m_ptr = *m_bucket_ptr + BucketSize + 1)

That is, now you can "conditionally" do the following:
Farebl::deque_dimensional<int, 7> container;
container.push_back(1);
container.push_back(2);
auto it = container.begin();
... // do (--it;) 1000 times;
... // do (++it;) 1000 times;
cout<<*it; // correct output: 1
++it;
cout<<*it; // correct output: 2